# pkg/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerScopes = "Bearer.Scopes"
)

// JWKSetKey jwk set key
type JWKSetKey struct {
	// Alg 署名検証アルゴリズム
	Alg string `json:"alg"`

	// E exponent 公開鍵を復元するための指数値
	E string `json:"e"`

	// Kid 鍵識別子
	Kid openapi_types.UUID `json:"kid"`

	// Kty RSAやEC等の暗号アルゴリズムファミリー
	Kty string `json:"kty"`

	// N modulus 公開鍵を復元するための公開鍵の絶対値
	N string `json:"n"`

	// Use 公開鍵の用途
	Use string `json:"use"`
}

// JWKSetSchema https://openid-foundation-japan.github.io/rfc7517.ja.html#anchor5
type JWKSetSchema struct {
	Keys []JWKSetKey `json:"keys"`
}

// PostCertsJSONRequestBody defines body for PostCerts for application/json ContentType.
type PostCertsJSONRequestBody = JWKSetKey